requirements.txt
gradio
gradio[oauth]

requests
python-dotenv
langgraph
langchain
langchain[openai]
langchain_community
langchain_experimental
langchain-mcp-adapters
langsmith

duckduckgo-search
openpyxl
youtube_transcript_api
pytube
tavily-python
assemblyai

mcp[cli]